+++ src/pages/_document.tsx
@@ -1,0 +1,14 @@
+/* eslint-disable check-file/filename-naming-convention */
+import { Html, Head, Main, NextScript } from "next/document";
+
+export default function Document() {
+  return (
+    <Html lang="es">
+      <Head />
+      <body>
+        <Main />
+        <NextScript />
+      </body>
+    </Html>
+  );
+}
